refactor(ask,think): share ollama streaming
